Trait ruma_common::api::EndpointError

pub trait EndpointError: OutgoingResponse + StdError + Sized + Send + 'static {
    // Required method
    fn from_http_response<T: AsRef<[u8]>>(response: Response<T>) -> Self;
}

Gives users the ability to define their own serializable / deserializable errors.

Required Methods

fn from_http_response<T: AsRef<[u8]>>(response: Response<T>) -> Self

Tries to construct Self from an http::Response.

This will always return Err variant when no error field is defined in the ruma_api macro.

Object Safety

This trait is not object safe.