Module ruma_common::directory

Common types for room directory endpoints.

Structs

Filter

A filter for public rooms lists

IncomingFilter

‘Incoming’ variant of Filter.

PublicRoomsChunk

A chunk of a room list response, describing one room.

PublicRoomsChunkInit

Initial set of mandatory fields of PublicRoomsChunk.

Enums

IncomingRoomNetwork

‘Incoming’ variant of RoomNetwork.

PublicRoomJoinRule

The rule used for users wishing to join a public room.

RoomNetwork

Information about which networks/protocols from application services on the homeserver from which to request rooms.