Enum ruma_common::events::SyncStateEvent

pub enum SyncStateEvent<C: StateEventContent + RedactContent> where
    C::Redacted: StateEventContent + RedactedEventContent,  {
    Original(OriginalSyncStateEvent<C>),
    Redacted(RedactedSyncStateEvent<C::Redacted>),
}

A possibly-redacted state event without a room_id.

SyncStateEvent implements the comparison traits using only the event_id field, a sorted list would be sorted lexicographically based on the event’s EventId.

Variants

Original(OriginalSyncStateEvent<C>)

Original, unredacted form of the event.

Redacted(RedactedSyncStateEvent<C::Redacted>)

Redacted form of the event with minimal fields.

Implementations

impl<C> SyncStateEvent<C> where
C: StateEventContent + RedactContent,
C::Redacted: StateEventContent + RedactedEventContent,
C::Redacted: StateEventContent<StateKey = C::StateKey>,

pub fn event_type(&self) -> StateEventType

Returns the type of this event.

pub fn event_id(&self) -> &EventId

Returns this event’s event_id field.

pub fn sender(&self) -> &UserId

Returns this event’s sender field.

pub fn origin_server_ts(&self) -> MilliSecondsSinceUnixEpoch

Returns this event’s origin_server_ts field.

pub fn state_key(&self) -> &C::StateKey

Returns this event’s state_key field.

pub fn as_original(&self) -> Option<&OriginalSyncStateEvent<C>>

Get the inner OriginalSyncStateEvent if this is an unredacted event.

pub fn into_full_event(self, room_id: OwnedRoomId) -> StateEvent<C>

Convert this sync event into a full event (one with a room_id field).

impl SyncStateEvent<RoomGuestAccessEventContent>

pub fn guest_access(&self) -> &GuestAccess

Obtain the guest access policy, regardless of whether this event is redacted.

impl SyncStateEvent<RoomHistoryVisibilityEventContent>

pub fn history_visibility(&self) -> &HistoryVisibility

Obtain the history visibility, regardless of whether this event is redacted.

impl SyncStateEvent<RoomJoinRulesEventContent>

pub fn join_rule(&self) -> &JoinRule

Obtain the join rule, regardless of whether this event is redacted.

impl SyncStateEvent<RoomMemberEventContent>

pub fn membership(&self) -> &MembershipState

Obtain the membership state, regardless of whether this event is redacted.

impl SyncStateEvent<RoomPowerLevelsEventContent>

pub fn power_levels(&self) -> RoomPowerLevels

Obtain the effective power levels, regardless of whether this event is redacted.