Module ruma_common::events::relation

This is supported on crate features events and unstable-msc2675 only.

Types describing event relations after MSC 2674, 2675, 2676, 2677.

Structs

AnnotationChunkunstable-msc2677

The first chunk of annotations with a token for loading more.

BundledReactionunstable-msc2677

Summary of all reactions with the given key to an event.

BundledReplacementunstable-msc2676

A bundled replacement.

BundledThreadunstable-msc3440

A bundled thread.

Relations

Precompiled list of relations to this event grouped by relation type.

Enums

BundledAnnotationunstable-msc2677

Type of bundled annotation.