Module ruma_common::events::room::message

Types for the m.room.message event.

Modules

feedback

Types for the m.room.message.feedback event.

Structs

AudioInfo

Metadata about an audio clip.

AudioMessageEventContent

The payload for an audio message.

EmoteMessageEventContent

The payload for an emote message.

FileInfo

Metadata about a file.

FileMessageEventContent

The payload for a file message.

FormattedBody

Common message event content fields for message types that have separate plain-text and formatted representations.

ImageMessageEventContent

The payload for an image message.

InReplyTo

Information about the event a “rich reply” is replying to.

KeyVerificationRequestEventContent

The payload for a key verification request message.

LocationInfo

Thumbnail info associated with a location.

LocationMessageEventContent

The payload for a location message.

NoticeMessageEventContent

The payload for a notice message.

RedactedRoomMessageEventContent

Redacted form of RoomMessageEventContent

Replacement

The event this relation belongs to replaces another event.

RoomMessageEventContent

The content of an m.room.message event.

ServerNoticeMessageEventContent

The payload for a server notice message.

TextMessageEventContent

The payload for a text message.

VideoInfo

Metadata about a video.

VideoMessageEventContent

The payload for a video message.

Enums

LimitType

Types of usage limits.

MessageFormat

The format for the formatted representation of a message body.

MessageType

The content that is specific to each message type variant.

Relation

Message event relationship.

ServerNoticeType

Types of server notices.

Type Definitions

OriginalRoomMessageEvent

An m.room.message event.

OriginalSyncRoomMessageEvent

An m.room.message event from a sync_events response.

RedactedRoomMessageEvent

An m.room.message event that has been redacted.

RedactedSyncRoomMessageEvent

An m.room.message event from a sync_events response that has been redacted.

RoomMessageEvent

An m.room.message event.

SyncRoomMessageEvent

An m.room.message event from a sync_events response.