Enum ruma_common::serde::CanonicalJsonValue

pub enum CanonicalJsonValue {
    Null,
    Bool(bool),
    Integer(Int),
    String(String),
    Array(Vec<CanonicalJsonValue>),
    Object(Object),
}

Represents a canonical JSON value as per the Matrix specification.

Variants

Null

Represents a JSON null value.

let v: CanonicalJsonValue = json!(null).try_into().unwrap();

Bool(bool)

Represents a JSON boolean.

let v: CanonicalJsonValue = json!(true).try_into().unwrap();

Integer(Int)

Represents a JSON integer.

let v: CanonicalJsonValue = json!(12).try_into().unwrap();

String(String)

Represents a JSON string.

let v: CanonicalJsonValue = json!("a string").try_into().unwrap();

Array(Vec<CanonicalJsonValue>)

Represents a JSON array.

let v: CanonicalJsonValue = json!(["an", "array"]).try_into().unwrap();

Object(Object)

Represents a JSON object.

The map is backed by a BTreeMap to guarantee the sorting of keys.

let v: CanonicalJsonValue = json!({ "an": "object" }).try_into().unwrap();

Implementations

impl CanonicalJsonValue

pub fn as_bool(&self) -> Option<bool>

If the CanonicalJsonValue is a Bool, return the inner value.

pub fn as_integer(&self) -> Option<Int>

If the CanonicalJsonValue is an Integer, return the inner value.

pub fn as_str(&self) -> Option<&str>

If the CanonicalJsonValue is a String, return a reference to the inner value.

pub fn as_array(&self) -> Option<&[CanonicalJsonValue]>

If the CanonicalJsonValue is an Array, return a reference to the inner value.

pub fn as_object(&self) -> Option<&Object>

If the CanonicalJsonValue is an Object, return a reference to the inner value.

pub fn as_array_mut(&mut self) -> Option<&mut Vec<CanonicalJsonValue>>

If the CanonicalJsonValue is an Array, return a mutable reference to the inner value.

pub fn as_object_mut(&mut self) -> Option<&mut Object>

If the CanonicalJsonValue is an Object, return a mutable reference to the inner value.

pub fn is_bool(&self) -> bool

Returns true if the CanonicalJsonValue is a Bool.

pub fn is_integer(&self) -> bool

Returns true if the CanonicalJsonValue is an Integer.

pub fn is_string(&self) -> bool

Returns true if the CanonicalJsonValue is a String.

pub fn is_array(&self) -> bool

Returns true if the CanonicalJsonValue is an Array.

pub fn is_object(&self) -> bool

Returns true if the CanonicalJsonValue is an Object.