Crate ruma_common

Common types for other ruma crates.

Modules

authentication

Common types for authentication.

directory

Common types for room directory endpoints.

encryption

Common types for encryption related tasks.

power_levels

Common types for the m.room.power_levels event.

presence

Common types for the presence module.

push

Common types for the push notifications module.

receipt

Common types for receipts.

thirdparty

Common types for the third party networks module.

to_device

Common types for the Send-To-Device Messaging

Structs

MilliSecondsSinceUnixEpoch

A timestamp represented as the number of milliseconds since the unix epoch.

SecondsSinceUnixEpoch

A timestamp represented as the number of seconds since the unix epoch.