Module ruma_common::canonical_json

Available on crate feature canonical-json only.

Canonical JSON types and related functions.

Enums

CanonicalJsonError

The set of possible errors when serializing to canonical JSON.

CanonicalJsonValue

Represents a canonical JSON value as per the Matrix specification.

JsonType

A JSON type enum for RedactionError variants.

RedactionError

Errors that can happen in redaction.

Functions

redact

Redacts an event using the rules specified in the Matrix client-server specification.

redact_content_in_place

Redacts event content using the rules specified in the Matrix client-server specification.

redact_in_place

Redacts an event using the rules specified in the Matrix client-server specification.

to_canonical_value

Fallible conversion from any value that impl’s Serialize to a CanonicalJsonValue.

try_from_json_map

Fallible conversion from a serde_json::Map to a CanonicalJsonObject.

Type Definitions

CanonicalJsonObject

The inner type of CanonicalJsonValue::Object.